Hello world!
This is a working Sinatra app, running with puma. It can be
containerized to run in production, and it can be containerized to run
unit tests, linting, and feature specifications. The website itself just
says hello.

There are currently 0 unit tests and 0 features, and the linter says
everything conforms to standard ruby.