Set up cucumber with capybara, firefox, and selenium
There still aren't any running tests, but the feature tests now open a
production build of the app in a browser, which they can navigate using
capybara.

Also, the additional system requirements seemed like a good reason to
add another build phase to the dockerfile so that unit tests and linting
don't waste time installing firefox or any of the related ruby
libraries.