Start linting yaml files
I'm not recommending that everyone start linting their yaml files. I
personally find yaml easier to read when the files are formatted
consistently (indents, quoted strings, etc.) so I don't want to
accidentally lapse and forget to quote something or indent an array.

It's important to me that I be consistent, so why not enforce it?